gpo: Script ext should not crash if script missing
If a user has manually removed a script, the
extension should not crash in an unapply removing
it.

+++ python/samba/gp_scripts_ext.py
@@ -39,7 +39,8 @@ class gp_scripts_ext(gp_pol_ext):
             self.gp_db.set_guid(guid)
             if str(self) in settings:
                 for attribute, script in settings[str(self)].items():
-                    os.unlink(script)
+                    if os.path.exists(script):
+                        os.unlink(script)
                     self.gp_db.delete(str(self), attribute)
             self.gp_db.commit()
